fix(formats): 19 Linksschiebungen, die in C undefiniert sind (MF-594)
R3 des Release-Ziels. Beide UBSan-Befunde im Volllauf: 2 → 0.

── Der Fund ────────────────────────────────────────────────────────────

Der Fuzzer ueber den echten Oeffnungspfad, unter UBSan im CI:

    src/formats/apridisk/uft_apridisk.c:22:54: runtime error:
        left shift of 173 by 24 places cannot be represented in type 'int'

    static uint32_t read_le32(const uint8_t *p) {
        return p[0] | (p[1] << 8) | (p[2] << 16) | (p[3] << 24);
    }

`p[3]` ist `uint8_t` und wird vor der Schiebung zu `int` befoerdert. Ab
128 passt das Ergebnis nicht in einen 32-Bit-`int`. Das ist in C nicht
haesslich, sondern **undefiniert** — und kein Uebersetzer sagt etwas.

Bei einem Format-Parser heisst das: das Byte, das die Regel bricht, ist
genau das Byte, das eine beschaedigte oder feindliche Datei liefert.

── Ein Sanitizer, den es hier bisher nicht gab ─────────────────────────

Dieser Baum hat unter MinGW keine Sanitizer — ASan/UBSan laufen nur im
CI unter Linux. Das gilt aber nicht fuer alles:

    gcc -fsanitize=shift -fsanitize-undefined-trap-on-error

braucht keine Laufzeitbibliothek, sondern setzt ein `ud2`. Damit laesst
sich die ganze Suite lokal bauen und laufen — und sie brach an **genau
den zwei Tests** ab, die auch der CI-UBSan-Lauf meldet:

    6 - test_air_cross_validate (ILLEGAL)
    87 - test_disk_open_fuzz (Failed)

Also ein Rotbeweis auf diesem Rechner, vor jeder Aenderung. Rezept fuer
das naechste Mal:

    cmake -S . -B build-shift -G "MinGW Makefiles" \
      -DCMAKE_C_FLAGS="-fsanitize=shift -fsanitize-undefined-trap-on-error -g -O0" \
      -DCMAKE_CXX_FLAGS="...dasselbe..."

Die Fundstelle holt `gdb -batch` aus dem Trap-Rahmen; `trap-on-error`
gibt keine Meldung aus.

── Was geaendert wurde ─────────────────────────────────────────────────

Acht Stellen im Produktionscode, alle dieselbe Bauart:

    src/formats/apridisk/uft_apridisk.c:22          read_le32
    src/formats/ldbs/uft_ldbs.c:100                 read_le32
    src/formats/qrst/uft_qrst.c:25                  read_le32
    src/formats/kfx/uft_kfstream_air.c:157          kf_extract_int
    src/algorithms/advanced/uft_bayesian_detect_v2.c:277
    src/fileops/uft_file_ops_extended.c:450
    src/protection/uft_atarist_macrodos.c:319
    src/protection/uft_c64_protection_enhanced.c:345

`kf_extract_int` ist der einzige Sonderfall: er castet auf `(int)` und
soll auch vorzeichenbehaftet herauskommen. Jetzt wird vorzeichenlos
zusammengesetzt und dann umgedeutet — implementierungsdefiniert statt
undefiniert, auf Zweierkomplement dasselbe Bitmuster wie gemeint.

Elf Zeilen im Testcode, in `test_air_cross_validate.c` (neun) und
`tests/flux_gen/fluxengine/flux_gen.c` (zwei).

Dass die letzte Stelle im TEST lag und nicht in `src/`, hat mein
Prueflauf zuerst nicht gesehen — er las nur `src/`. Gefunden hat sie erst
gdb.

── Was NICHT geaendert wurde ───────────────────────────────────────────

`src/samdisk/` und `src/a8rawconv/` tragen dieselbe Bauart an dreizehn
Stellen. Sie sind **nicht im Bau** (`grep -c "src/samdisk/"
UnifiedFloppyTool.pro` -> 0) — eingekaufte Referenzquellen, kein
Produktionscode. Unberuehrt gelassen und hier benannt, statt still
mitgezaehlt.

Ein statisches Tor gegen diese Klasse waere eine Fehlalarm-Maschine: die
verbleibenden Treffer meines Pruefers sind `uint32_t`-Rotationen
(`sum >> 1 | sum << 31`) und `1ULL << 40` — harmlos, aber ohne
Typwissen nicht von echten Faellen zu trennen. Der belastbare Prueftisch
ist der dynamische oben.

── Messung ─────────────────────────────────────────────────────────────

    Suite unter -fsanitize=shift  vorher: 264/266   nachher: 266/266
    Normalbau qmake + cmake       gruen,  266/266
    check_consistency.py          alle Tore 0

Der Effekt im CI-UBSan-Volllauf steht nach dem naechsten Lauf fest.

Nebenbei zurueckgenommen: `OPEN_ITEMS.md` P0-11 behauptete, der
Leck-Rueckstand sei reiner Testcode und der Produktionspfad belegt
sauber. MF-592 hat das widerlegt. Der Eintrag sagt jetzt, was gemessen
ist.

+++ src/algorithms/advanced/uft_bayesian_detect_v2.c
@@ -274,8 +274,10 @@ static bool validate_adf(const uint8_t* data, size_t len) {
     // Bootblock checksum (simplified check)
     uint32_t checksum = 0;
     for (int i = 0; i < 1024; i += 4) {
-        uint32_t val = (data[i] << 24) | (data[i+1] << 16) | 
-                       (data[i+2] << 8) | data[i+3];
+        /* MF-594: ohne Cast ist `data[i] << 24` ab 128 undefiniert. */
+        uint32_t val = ((uint32_t)data[i]   << 24) |
+                       ((uint32_t)data[i+1] << 16) |
+                       ((uint32_t)data[i+2] <<  8) | (uint32_t)data[i+3];
         checksum += val;
     }
